Moved all instances of get-zealot into the api
