+++ src/app/core/pools/create-pool.ts
@@ -1,4 +1,3 @@
-import {getZealot} from "src/js/flows/getZealot"
 import Current from "src/js/state/Current"
 import {Thunk} from "src/js/state/types"
 import {syncPool} from "./sync-pool"
@@ -9,8 +8,8 @@ type Props = {
 
 export const createPool =
   ({name}: Props): Thunk<Promise<string>> =>
-  async (dispatch, getState) => {
-    const zealot = await dispatch(getZealot())
+  async (dispatch, getState, {api}) => {
+    const zealot = await api.getZealot()
     const lakeId = Current.getLakeId(getState())
     return zealot.createPool(name).then((res) => {
       return dispatch(syncPool(res.pool.id, lakeId)).then(() => res.pool.id)
