+++ src/js/api/index.ts
@@ -1,23 +1,23 @@
 import {Abortables} from "src/app/core/models/abortables"
 import toast from "react-hot-toast"
-import {getZealot} from "../flows/getZealot"
+import {getZealot} from "./core/get-zealot"
 import {AppDispatch, State} from "../state/types"
-import {StorageApi} from "./storage"
-import {ConfigsApi, ToolbarApi} from "./ui-apis"
 import {QueriesApi} from "./queries/queries-api"
 import {PoolsApi} from "./pools/pools-api"
-import {getPath, PathName} from "./paths/get-path"
+import {getPath, PathName} from "./core/get-path"
 import {CommandsApi} from "./commands/cmmands-api"
 import {LoadersApi} from "./loaders/loaders-api"
 import {Detail, MenusApi, Search} from "./menus/menus-api"
+import {ConfigurationsApi} from "./configurations/configurations-api"
+import {ToolbarsApi} from "./toolbars/toolbars-api"
+import {BrimLake} from "../brim"
 
 export default class BrimApi {
   public abortables = new Abortables()
   public commands = new CommandsApi()
   public loaders = new LoadersApi()
-  public toolbar: ToolbarApi
-  public configs: ConfigsApi
-  public storage: StorageApi
+  public toolbar: ToolbarsApi
+  public configs: ConfigurationsApi
   public queries: QueriesApi
   public pools: PoolsApi
   public toast = toast
@@ -27,20 +27,17 @@ export default class BrimApi {
   }
 
   private dispatch: AppDispatch
-  private getState: () => State
 
   init(d: AppDispatch, gs: () => State) {
     this.dispatch = d
-    this.getState = gs
-    this.toolbar = new ToolbarApi(d, gs)
-    this.configs = new ConfigsApi(d, gs)
-    this.storage = new StorageApi(d, gs)
+    this.toolbar = new ToolbarsApi(d, gs)
+    this.configs = new ConfigurationsApi(d, gs)
     this.queries = new QueriesApi(d)
     this.pools = new PoolsApi(d)
   }
 
-  getZealot() {
-    return this.dispatch(getZealot())
+  getZealot(lake?: BrimLake, env?: "node" | "web") {
+    return this.dispatch(getZealot(lake, env))
   }
 
   getPath(name: PathName) {
