+++ src/js/flows/exportResults.ts
@@ -7,7 +7,6 @@ import brim from "../brim"
 import Columns from "../state/Columns"
 import Results from "../state/Results"
 import {Thunk} from "../state/types"
-import {getZealot} from "./getZealot"
 
 const streamPipeline = util.promisify(pipeline)
 
@@ -33,8 +32,8 @@ export default (
     filePath: string,
     format: ResponseFormat
   ): Thunk<Promise<string>> =>
-  async (dispatch, getState): Promise<string> => {
-    const zealot = await dispatch(getZealot(undefined, "node"))
+  async (dispatch, getState, {api}): Promise<string> => {
+    const zealot = await api.getZealot(undefined, "node")
     const columns = Columns.getCurrentTableColumns(getState())
     const originalQuery = Results.getQuery(getState())
     const exportQuery = prepareProgram(format, originalQuery, columns)
