+++ src/js/flows/inspectSearch.ts
@@ -2,12 +2,12 @@ import Current from "../state/Current"
 import SearchBar from "../state/SearchBar"
 import Tab from "../state/Tab"
 import {Thunk} from "../state/types"
-import {getZealot} from "./getZealot"
 import {annotateQuery} from "./search/mod"
 
 export const inspectSearch =
-  (): Thunk<Promise<string>> => async (dispatch, getState) => {
-    const zealot = await dispatch(getZealot())
+  (): Thunk<Promise<string>> =>
+  async (dispatch, getState, {api}) => {
+    const zealot = await api.getZealot()
     const program = SearchBar.getSearchProgram(getState())
     const [from, to] = Tab.getSpan(getState())
     const poolId = Current.getPoolId(getState())
