+++ src/js/flows/renamePool.ts
@@ -1,10 +1,9 @@
 import {syncPool} from "src/app/core/pools/sync-pool"
 import {Thunk} from "../state/types"
-import {getZealot} from "./getZealot"
 
 export default (poolId: string, name: string): Thunk<Promise<void>> =>
-  async (dispatch) => {
-    const zealot = await dispatch(getZealot())
+  async (dispatch, gs, {api}) => {
+    const zealot = await api.getZealot()
     await zealot.updatePool(poolId, {name})
     await dispatch(syncPool(poolId))
   }
