+++ src/js/flows/subscribeEvents.ts
@@ -1,11 +1,10 @@
 import {BrimLake} from "../brim"
 import {Thunk} from "../state/types"
-import {getZealot} from "./getZealot"
 
 export const subscribeEvents =
   (lake?: BrimLake): Thunk<Promise<EventSource>> =>
-  async (dispatch) => {
-    const zealot = await dispatch(getZealot(lake))
+  async (d, gs, {api}) => {
+    const zealot = await api.getZealot(lake)
 
     return zealot.subscribe()
   }
